requirements.txt
requests==2.28.1
ruamel.base==1.0.0
semantic_version==2.10.0
typer==0.6.1